+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 #### Fixes
 
-* Your contribution here.
+* [#681](https://github.com/ruby-grape/grape-swagger/pull/681): Provide error schemas when an endpoint can return a 204 - [@adstratm](https://github.com/adstratm).
+* [#683](https://github.com/ruby-grape/grape-swagger/pull/683): Fix handling of arrays of complex entities in params so that valid OpenAPI spec is generated - [@jdmurphy](https://github.com/jdmurphy).
 
 ###  0.29.0 (May 22, 2018)