+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 #### Features
 
-* [#684](https://github.com/ruby-grape/grape-swagger/pull/684): Add response headers - [@jdmurphy](https://github.com/jdmurphy).
 * Your contribution here.
+
+#### Fixes
+
+* Your contribution here.
+
+###  0.30.0 (July 19, 2018)
+
+#### Features
+
+* [#684](https://github.com/ruby-grape/grape-swagger/pull/684): Add response headers - [@jdmurphy](https://github.com/jdmurphy).
 
 #### Fixes